# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://api-task-management.herokuapp.com/",
            "description": "Api pour gestion de taches"
        }
    ],
    "paths": {
        "/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "operationId": "App\\Http\\Controllers\\TaskController::index",
                "parameters": [
                    {
                        "name": "completed",
                        "in": "path",
                        "description": "Les task completed",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task index",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unauthorised.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "default": "unauthorised"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "tasks"
                ],
                "operationId": "App\\Http\\Controllers\\TaskController::store",
                "parameters": [
                    {
                        "name": "body",
                        "in": "query",
                        "description": "Body de la task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "completed",
                        "in": "query",
                        "description": "Task fini ? oui/non",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unauthorised.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "default": "unauthorised"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "operationId": "App\\Http\\Controllers\\TaskController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "task not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "default": "task not found"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unauthorised.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "default": "unauthorised"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "tasks"
                ],
                "operationId": "App\\Http\\Controllers\\TaskController::update",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "ID de la task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "description": "Body de la task",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "completed",
                        "in": "query",
                        "description": "Task fini ? oui/non",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "task not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "default": "task not found"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unauthorised.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "default": "unauthorised"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tasks"
                ],
                "operationId": "App\\Http\\Controllers\\TaskController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete task"
                    },
                    "404": {
                        "description": "task not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "default": "task not found"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unauthorised.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "default": "unauthorised"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "users"
                ],
                "operationId": "App\\Http\\Controllers\\UserController::register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "users"
                ],
                "operationId": "App\\Http\\Controllers\\UserController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Incorrect login or password.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "default": "incorrect login or password."
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "description": "task",
                "properties": {
                    "id": {
                        "type": "integer",
                        "default": "1"
                    },
                    "user_id": {
                        "type": "integer",
                        "default": "1"
                    },
                    "body": {
                        "type": "string",
                        "default": "some text"
                    },
                    "completed": {
                        "type": "boolean",
                        "default": "0"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "description": "user",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "remember_token": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}